Change the way sim-tests are run.  Abandon running them using cargo test, add a function that runs them via python import.  We'll build a wheel and install it.
