+++ Cargo.toml
@@ -34,7 +34,7 @@ tracing = "0.1"
 ctor = "0.2.8"
 log = "0.4.22"
 exec = { version = "0.3.1", optional = true }
-pyo3 = { version = "0.25.1", features = ["auto-initialize"], optional = true }
+pyo3 = { version = "0.25.1", features = ["auto-initialize","extension-module"], optional = true }
 tracing-core = "0.1.34"
 tracing-subscriber = { version = "0.3", features = ["alloc"] }
 simplelog = "0.12.2"
