+++ src/simulator/mod.rs
@@ -1,3 +1,5 @@
+pub mod tests;
+
 use std::cell::RefCell;
 
 use clvm_traits::{ClvmEncoder, ToClvm};
@@ -23,6 +25,9 @@ use crate::common::types::{
     GetCoinStringParts, Hash, IntoErr, Node, Program, Puzzle, PuzzleHash, Sha256tree, Spend,
     ToQuotedProgram,
 };
+use crate::simulator::tests::potato_handler_sim::test_funs as potato_handler_sim_tests;
+use crate::simulator::tests::simenv::test_funs as simenv_tests;
+use crate::test_support::calpoker::test_funs as calpoker_tests;
 
 #[derive(Debug, Clone)]
 pub struct IncludeTransactionResult {
@@ -602,4 +607,26 @@ impl Simulator {
             &amount,
         ))
     }
+}
+
+#[pyfunction]
+fn run_simulation_tests() {
+    let ref_lists = [
+        &simenv_tests(),
+        &calpoker_tests(),
+        &potato_handler_sim_tests(),
+    ];
+    for test_set in ref_lists.iter() {
+        for (name, f) in test_set.iter() {
+            eprintln!("{} ...", name);
+            f();
+            eprintln!("{} ... ok\n", name);
+        }
+    }
+}
+
+#[pymodule]
+fn chia_gaming(_py: Python, m: Bound<'_, PyModule>) -> PyResult<()> {
+    m.add_function(wrap_pyfunction!(run_simulation_tests, &m)?)?;
+    Ok(())
 }
