+++ src/simulator/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod potato_handler_sim;
+pub mod simenv;
