+++ src/test_support/mod.rs
@@ -1,0 +1,5 @@
+pub mod calpoker;
+#[cfg(any(test, feature = "sim-tests", feature = "simulator"))]
+pub mod debug_game;
+pub mod game;
+pub mod peer;
