+++ src/test_support/peer/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(any(test, feature = "sim-tests", feature = "simulator"))]
+pub mod potato_handler;
