--- src/tests/peer/mod.rs
@@ -1,3 +1,0 @@
-pub mod potato_handler;
-#[cfg(feature = "sim-tests")]
-pub mod potato_handler_sim;
